refactor: more consistent and flexible precision trait usage

+++ server/model/src/api/point_struct.rs
@@ -82,7 +82,7 @@ impl ToCollection for PointStruct {
 
 impl ToText for PointStruct {
     fn to_text(self, sep_1: &str, sep_2: &str, _poly_sep: bool) -> String {
-        format!("{}{}{}{}", self.lat as f32, sep_1, self.lon as f32, sep_2)
+        format!("{}{}{}{}", self.lat, sep_1, self.lon, sep_2)
     }
 }
